+++ src/main/java/com/program/commandLine/customer/Customer.java
@@ -1,0 +1,10 @@
+package com.program.commandLine.customer;
+
+import java.util.UUID;
+
+public interface Customer {
+
+    UUID getCustomerId();
+
+    String getName();
+}
