+++ src/main/java/com/program/commandLine/io/Input.java
@@ -1,0 +1,7 @@
+package com.program.commandLine.io;
+
+public interface Input {
+    String input();
+
+    String input(String message);
+}
