+++ src/main/java/com/program/commandLine/voucher/Voucher.java
@@ -1,0 +1,11 @@
+package com.program.commandLine.voucher;
+
+import java.util.UUID;
+
+public interface Voucher {
+    UUID getVoucherId();
+    String getVoucherType();
+    int getVoucherDiscount();
+    int discountPrice(int beforeDiscount);
+
+}
