--- src/main/java/com/program/voucher/io/Input.java
@@ -1,7 +1,0 @@
-package com.program.voucher.io;
-
-public interface Input {
-    String input();
-
-    String input(String message);
-}
