--- src/main/java/com/program/voucher/model/Voucher.java
@@ -1,11 +1,0 @@
-package com.program.voucher.model;
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getVoucherId();
-    String getVoucherType();
-    int getVoucherDiscount();
-    int discountPrice(int beforeDiscount);
-
-}
