--- src/main/java/com/program/voucher/repository/MemoryVoucherRepository.java
@@ -1,38 +1,0 @@
-package com.program.voucher.repository;
-
-import com.program.voucher.model.Voucher;
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-
-
-@Component(value = "voucherRepository")
-@Profile("local")
-public class MemoryVoucherRepository implements VoucherRepository {
-
-    private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>(); // hashmap ? map ?
-
-    @Override
-    public Optional<Voucher> findById(UUID voucherId) {
-        return Optional.ofNullable(storage.get(voucherId));
-    }
-
-    @Override
-    public Voucher insertVoucher(Voucher voucher) {
-        storage.put(voucher.getVoucherId(), voucher);
-        return voucher;
-    }
-
-    @Override
-    public List<Voucher> findAll() {
-        return storage.values().stream().toList();
-    }
-
-
-}
