+++ src/test/java/com/program/commandLine/repository/FileVoucherRepositoryTest.java
@@ -1,0 +1,28 @@
+package com.program.commandLine.repository;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class FileVoucherRepositoryTest {
+
+    @BeforeEach
+    void setUp() {
+    }
+
+    @AfterEach
+    void tearDown() {
+    }
+
+    @Test
+    void findById() {
+    }
+
+    @Test
+    void insertVoucher() {
+    }
+
+    @Test
+    void findAll() {
+    }
+}
