+++ src/test/java/com/program/commandLine/repository/MemoryVoucherRepositoryTest.java
@@ -1,0 +1,18 @@
+package com.program.commandLine.repository;
+
+import org.junit.jupiter.api.Test;
+
+class MemoryVoucherRepositoryTest {
+
+    @Test
+    void findById() {
+    }
+
+    @Test
+    void insertVoucher() {
+    }
+
+    @Test
+    void findAll() {
+    }
+}
