--- src/test/java/com/program/voucher/repository/FileVoucherRepositoryTest.java
@@ -1,30 +1,0 @@
-package com.program.voucher.repository;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class FileVoucherRepositoryTest {
-
-    @BeforeEach
-    void setUp() {
-    }
-
-    @AfterEach
-    void tearDown() {
-    }
-
-    @Test
-    void findById() {
-    }
-
-    @Test
-    void insertVoucher() {
-    }
-
-    @Test
-    void findAll() {
-    }
-}
