--- src/test/java/com/program/voucher/repository/MemoryVoucherRepositoryTest.java
@@ -1,20 +1,0 @@
-package com.program.voucher.repository;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class MemoryVoucherRepositoryTest {
-
-    @Test
-    void findById() {
-    }
-
-    @Test
-    void insertVoucher() {
-    }
-
-    @Test
-    void findAll() {
-    }
-}
